Rename model names (#281)
