+++ armory/baseline_models/keras/cifar.py
