+++ armory/baseline_models/keras/inception_resnet_v2.py
