+++ armory/baseline_models/keras/mnist.py
