+++ armory/baseline_models/keras/resisc45.py
