+++ armory/baseline_models/keras/resnet50.py
