+++ armory/baseline_models/pytorch/cifar.py
